+++ cmd/s3-authmate/main.go
@@ -35,6 +35,7 @@ const (
 	poolHealthcheckTimeout = 5 * time.Second
 	poolRebalanceInterval  = 30 * time.Second
 	poolStreamTimeout      = 10 * time.Second
+	minWaiterPollInterval  = 50 * time.Millisecond
 
 	// a month.
 	defaultLifetime = 30 * 24 * time.Hour
@@ -47,6 +48,7 @@ type PoolConfig struct {
 	HealthcheckTimeout time.Duration
 	StreamTimeout      time.Duration
 	RebalanceInterval  time.Duration
+	WaiterPollInterval time.Duration
 }
 
 var (
@@ -77,6 +79,8 @@ var (
 	poolHealthcheckTimeoutFlag time.Duration
 	poolRebalanceIntervalFlag  time.Duration
 	poolStreamTimeoutFlag      time.Duration
+
+	waiterPollIntervalFlag time.Duration
 )
 
 const (
@@ -279,6 +283,13 @@ It will be ceil rounded to the nearest amount of epoch.`,
 				Value:       poolRebalanceInterval,
 			},
 			&cli.DurationFlag{
+				Name:        "container-ops-poll-interval",
+				Usage:       `Polling interval for container operation waiter (50ms by default)`,
+				Required:    false,
+				Destination: &waiterPollIntervalFlag,
+				Value:       minWaiterPollInterval,
+			},
+			&cli.DurationFlag{
 				Name:        "pool-stream-timeout",
 				Usage:       `Timeout for individual operation in streaming RPC`,
 				Required:    false,
@@ -310,6 +321,7 @@ It will be ceil rounded to the nearest amount of epoch.`,
 				HealthcheckTimeout: poolHealthcheckTimeoutFlag,
 				StreamTimeout:      poolStreamTimeoutFlag,
 				RebalanceInterval:  poolRebalanceIntervalFlag,
+				WaiterPollInterval: max(waiterPollIntervalFlag, minWaiterPollInterval),
 			}
 
 			// authmate doesn't require anonKey for work, but let's create random one.
@@ -513,6 +525,13 @@ func obtainSecret() *cli.Command {
 				Value:       poolRebalanceInterval,
 			},
 			&cli.DurationFlag{
+				Name:        "container-ops-poll-interval",
+				Usage:       `Polling interval for container operation waiter (50ms by default)`,
+				Required:    false,
+				Destination: &waiterPollIntervalFlag,
+				Value:       minWaiterPollInterval,
+			},
+			&cli.DurationFlag{
 				Name:        "pool-stream-timeout",
 				Usage:       `Timeout for individual operation in streaming RPC`,
 				Required:    false,
@@ -539,6 +558,7 @@ func obtainSecret() *cli.Command {
 				HealthcheckTimeout: poolHealthcheckTimeoutFlag,
 				StreamTimeout:      poolStreamTimeoutFlag,
 				RebalanceInterval:  poolRebalanceIntervalFlag,
+				WaiterPollInterval: max(waiterPollIntervalFlag, minWaiterPollInterval),
 			}
 
 			// authmate doesn't require anonKey for work, but let's create random one.
@@ -618,6 +638,7 @@ func createNeoFS(ctx context.Context, log *zap.Logger, cfg PoolConfig, anonSigne
 		MaxObjectSize:        int64(ni.MaxObjectSize()),
 		IsSlicerEnabled:      isSlicerEnabled,
 		IsHomomorphicEnabled: !ni.HomomorphicHashingDisabled(),
+		WaiterPollInterval:   cfg.WaiterPollInterval,
 	}
 
 	neoFS := neofs.NewNeoFS(p, signer, anonSigner, neofsCfg, ni)